+++ src/tidop/viewer/renderer/Renderer.cpp
@@ -79,7 +79,7 @@ void Renderer::resize(int w, int h)
 	float phi = camera->getPhi();
 	float radius = camera->getRadius();
 
-	double aspectRatio = static_cast<double>(w) / h;
+	float aspectRatio = static_cast<float>(w) / h;
 	camera = TrackballCamera::perspectiveCamera(consts::grad_to_rad<float> * 45.0f, aspectRatio, 0.1, 10000);
 
 	camera->setTheta(theta);
